moved DatabaseObject to subdirectory database-values
