+++ test/scripts/database_values/index.js
@@ -1,0 +1,7 @@
+// jshint esversion: 8
+
+describe('database-values', () => {
+
+  require('./database_object');
+
+});
