+++ test/scripts/index.js
@@ -1,3 +1,3 @@
 require('./adapters');
 require('./database');
-require('./database_object');
+require('./database_values');
